Refactoring. Requests moved to dto/request, responses moved to dto/response
Response class has been created for GET request

Endpoint for get all print jobs has been created
