+++ Backend/.gitignore
@@ -69,6 +69,8 @@ gradle-app.setting
 .idea/**/usage.statistics.xml
 .idea/**/dictionaries
 .idea/**/shelf
+.idea/**/.name
+.idea/**/compiler.xml
 
 # Generated files
 .idea/**/contentModel.xml
