+++ Backend/src/main/java/idealab/api/dto/request/PrintJobUpdateRequest.java
@@ -1,4 +1,4 @@
-package idealab.api.dto;
+package idealab.api.dto.request;
 
 import static idealab.api.models.Status.*;
