+++ Backend/src/main/java/idealab/api/dto/response/GenericResponse.java
@@ -1,4 +1,4 @@
-package idealab.api.dto;
+package idealab.api.dto.response;
 
 public class GenericResponse {
